rename save methods to post_answer and post_question for clarity; update related interfaces and implementations

+++ managers/questions_related_managers/questions_manager.cpp
@@ -2,15 +2,30 @@
 
 questions_manager::questions_manager(unique_ptr<IQuestionRepository> repo) : questions_repo(std::move(repo)) {}
 
-void questions_manager::post_question(int from_user_id, const string &text , bool is_anonymous, int parent_question_id = -1)
+void questions_manager::post_question(int from_user_id, const string &text , bool is_anonymous, int parent_question_id)
 {
-    /*gotta fix up the questions and answers dao by changing the fact that it takes an object of each type
-    so it can save, which is incompatible due to the fact that i cant just build a quiestion by assigning its id
-    which gets me stuck into a loop where i cant make an object to call save and cant call save cuz i need an object*/
+    if(!validator::validate_text_size(text, validator::MAX_POST_LENGTH))
+    {
+        throw validation_exception("Your post is longer than 1000 characters");
+    }
+
+    questions_repo->post_question(from_user_id, text, is_anonymous, parent_question_id);
 }
 
-void questions_manager::delete_question(int question_id)
+void questions_manager::delete_question(int question_id, int user_id)
 {
+    unique_ptr<questions> question = questions_repo->find_by_id(question_id);
+
+    if(question == nullptr)
+    {
+        throw not_found_exception("This question doesn't exist");
+    }
+
+    if(question->from_user_id_getter() != user_id)
+    {
+        throw authentication_exception("This post isnt yours to delete");
+    }
+
     questions_repo->remove(question_id);
 }
 
@@ -26,6 +41,13 @@ vector<questions> questions_manager::get_user_questions(int from_user_id)
 
 unique_ptr<questions> questions_manager::get_question(int question_id)
 {
-    return questions_repo->find_by_id(question_id) ;
+    unique_ptr<questions> question = questions_repo->find_by_id(question_id) ;
+
+    if(question == nullptr)
+    {
+        throw not_found_exception("This question doesn't exist");
+    }
+
+    return question;
 }
